// src/api/types.gen.ts
export type BehaviorPolicyBehavior = {
    /**
     * The  arguments that the referenced validator type requires.
     */
    arguments?: {
        [key: string]: unknown;
    };
    /**
     * The unique identifier of a pre-defined behavior.
     */
    id: string;
};
// ...
export type BehaviorPolicyDeserialization = {
    publish?: BehaviorPolicyDeserializer;
    will?: BehaviorPolicyDeserializer;
};

/**
 * The matching rules the policy applies.
 */
export type BehaviorPolicyMatching = {
    /**
     * The regex pattern to match the client id against.
     */
    clientIdRegex: string;
};
// ...
export type BehaviorPolicyOnTransition = {
    'Connection.OnDisconnect'?: BehaviorPolicyOnEvent;
    'Event.OnAny'?: BehaviorPolicyOnEvent;
    'Mqtt.OnInboundConnect'?: BehaviorPolicyOnEvent;
    'Mqtt.OnInboundDisconnect'?: BehaviorPolicyOnEvent;
    'Mqtt.OnInboundPublish'?: BehaviorPolicyOnEvent;
    'Mqtt.OnInboundSubscribe'?: BehaviorPolicyOnEvent;
    /**
     * The exact state from which the transition happened. Alternatively a state filter can be used.
     */
    fromState: string;
    /**
     * The exact state to which the transition happened. Alternatively a state filter can be used.
     */
    toState: string;
};

/**
 * A policy which is used to validate and execute certain actions based on the validation result.
 */
export type BehaviorPolicy = {
    behavior: BehaviorPolicyBehavior;
    /**
     * The formatted UTC timestamp indicating when the policy was created.
     */
    readonly createdAt?: string;
    deserialization?: BehaviorPolicyDeserialization;
    /**
     * The unique identifier of the policy.
     */
    id: string;
    /**
     * The formatted UTC timestamp indicating when the policy was updated the last time.
     */
    readonly lastUpdatedAt?: string;
    matching: BehaviorPolicyMatching;
    onTransitions?: Array<BehaviorPolicyOnTransition>;
};
// ...
export type JsonNode = {
    [key: string]: unknown;
};
// ...
export type DataPolicyMatching = {
    /**
     * The topic filter for which the policy is matched.
     */
    topicFilter: string;
};

/**
 * One or more operations the outcome of the validation triggers.  When this field is empty, the outcome of the policy validation does not trigger any operations.
 */
export type DataPolicyAction = {
    /**
     * The pipeline to execute, when this action is triggered. The operations in the pipeline are executed in-order.
     */
    pipeline?: Array<PolicyOperation>;
};
// ...
export type DataPolicyValidation = {
    /**
     * The validators of the policy.
     */
    validators?: Array<DataPolicyValidator>;
};

/**
 * A data policy which is used to validate and execute certain actions based on the validation result.
 */
export type DataPolicy = {
    /**
     * The formatted UTC timestamp indicating when the policy was created.
     */
    readonly createdAt?: string;
    /**
     * The unique identifier of the policy.
     */
    id: string;
    /**
     * The formatted UTC timestamp indicating when the policy was updated the last time.
     */
    readonly lastUpdatedAt?: string;
    matching: DataPolicyMatching;
    onFailure?: DataPolicyAction;
    onSuccess?: DataPolicyAction;
    validation?: DataPolicyValidation;
};
// ...
export type PolicySchema = {
    /**
     * The schema type dependent arguments.
     */
    arguments?: {
        [key: string]: string;
    };
    /**
     * The formatted UTC timestamp when the schema was created.
     */
    readonly createdAt?: string;
    /**
     * The unique identifier of the schema.
     */
    id: string;
    /**
     * The base64 encoded schema definition.
     */
    schemaDefinition: string;
    /**
     * The type of the schema.
     */
    type: string;
    /**
     * The version of the schema.
     */
    readonly version?: number;
};

/**
 * A listing of schemas.
 */
export type SchemaList = {
    _links?: PaginationCursor;
    /**
     * List of result items that are returned by this endpoint
     */
    items?: Array<PolicySchema>;
};

export type Script = {
    /**
     * The formatted UTC timestamp when the script was created.
     */
    readonly createdAt?: string;
    /**
     * A string of free-form text describing the function.
     */
    description?: string;
    /**
     * The type of the function.
     */
    functionType: 'TRANSFORMATION';
    /**
     * The unique identifier of the script.
     */
    id: string;
    /**
     * The base64 encoded function source code.
     */
    source: string;
    /**
     * The version of the script.
     */
    readonly version?: number;
};
// ...
export type BridgeCustomUserProperty = {
    /**
     * The key the from the property
     */
    key: string;
    /**
     * The value the from the property
     */
    value: string;
};

/**
 * localSubscriptions associated with the bridge
 */
export type LocalBridgeSubscription = {
    /**
     * The customUserProperties for this subscription
     */
    customUserProperties?: Array<BridgeCustomUserProperty>;
    /**
     * The destination topic for this filter set.
     */
    destination: string;
    /**
     * The exclusion patterns
     */
    excludes?: Array<string | null> | null;
    /**
     * The filters for this subscription.
     */
    filters: Array<string>;
    /**
     * The maxQoS for this subscription.
     */
    maxQoS: 0 | 1 | 2;
    /**
     * The preserveRetain for this subscription
     */
    preserveRetain?: boolean;
    /**
     * The limit of this bridge for QoS-1 and QoS-2 messages.
     */
    queueLimit?: number | null;
};

/**
 * remoteSubscriptions associated with the bridge
 */
export type BridgeSubscription = {
    /**
     * The customUserProperties for this subscription
     */
    customUserProperties?: Array<BridgeCustomUserProperty>;
    /**
     * The destination topic for this filter set.
     */
    destination: string;
    /**
     * The filters for this subscription.
     */
    filters: Array<string>;
    /**
     * The maxQoS for this subscription.
     */
    maxQoS: 0 | 1 | 2;
    /**
     * The preserveRetain for this subscription
     */
    preserveRetain?: boolean;
};

/**
 * Information associated with the runtime of this adapter
 */
export type Status = {
    /**
     * A mandatory connection status field.
     */
    connection?: 'CONNECTED' | 'DISCONNECTED' | 'STATELESS' | 'UNKNOWN' | 'ERROR';
    /**
     * The identifier of the object
     */
    id?: string;
    /**
     * The datetime of the last activity through this connection
     */
    lastActivity?: string;
    /**
     * A message associated with the state of a connection
     */
    message?: string;
    /**
     * A object status field.
     */
    runtime?: 'STARTED' | 'STOPPED';
    /**
     * The datetime the object was 'started' in the system.
     */
    startedAt?: string;
    /**
     * The type of the object
     */
    type?: string;
};

/**
 * tlsConfiguration associated with the bridge
 */
export type TlsConfiguration = {
    /**
     * The cipherSuites from the config
     */
    cipherSuites?: Array<string>;
    /**
     * If TLS is used
     */
    enabled?: boolean;
    /**
     * The handshakeTimeout from the config
     */
    handshakeTimeout?: number;
    /**
     * The keystorePassword from the config
     */
    keystorePassword?: string;
    /**
     * The keystorePath from the config
     */
    keystorePath?: string | null;
    /**
     * The keystoreType from the config
     */
    keystoreType?: string;
    /**
     * The privateKeyPassword from the config
     */
    privateKeyPassword?: string;
    /**
     * The protocols from the config
     */
    protocols?: Array<string>;
    /**
     * The truststorePassword from the config
     */
    truststorePassword?: string;
    /**
     * The truststorePath from the config
     */
    truststorePath?: string | null;
    /**
     * The truststoreType from the config
     */
    truststoreType?: string;
    /**
     * The verifyHostname from the config
     */
    verifyHostname?: boolean;
} | null;

/**
 * websocketConfiguration associated with the bridge
 */
export type WebsocketConfiguration = {
    /**
     * If Websockets are used
     */
    enabled?: boolean;
    /**
     * The server path used by the bridge client. This must be setup as path at the remote broker
     */
    serverPath?: string;
    /**
     * The sub-protocol used by the bridge client. This must be supported by the remote broker
     */
    subProtocol?: string;
} | null;

export type Bridge = {
    /**
     * The cleanStart value associated the the MQTT connection.
     */
    cleanStart: boolean;
    /**
     * The client identifier associated the the MQTT connection.
     */
    clientId?: string | null;
    /**
     * The host the bridge connects to - a well formed hostname, ipv4 or ipv6 value.
     */
    host: string;
    /**
     * The bridge id, must be unique and only contain alpha numeric characters with spaces and hyphens.
     */
    id: string;
    /**
     * The keepAlive associated the the MQTT connection.
     */
    keepAlive: number;
    /**
     * localSubscriptions associated with the bridge
     */
    localSubscriptions?: Array<LocalBridgeSubscription>;
    /**
     * Is loop prevention enabled on the connection
     */
    loopPreventionEnabled?: boolean;
    /**
     * Loop prevention hop count
     */
    loopPreventionHopCount?: number;
    /**
     * The password value associated the the MQTT connection.
     */
    password?: string | null;
    /**
     * If this flag is set to true, any outgoing mqtt messages with QoS-1 or QoS-2 will be persisted on disc in case disc persistence is active.If this flag is set to false, the QoS of any outgoing mqtt messages will be set to QoS-0 and no traffic will be persisted on disc.
     */
    persist?: boolean | null;
    /**
     * The port number to connect to
     */
    port: number;
    /**
     * remoteSubscriptions associated with the bridge
     */
    remoteSubscriptions?: Array<BridgeSubscription>;
    /**
     * The sessionExpiry associated the the MQTT connection.
     */
    sessionExpiry: number;
    status?: Status;
    tlsConfiguration?: TlsConfiguration;
    /**
     * The username value associated the the MQTT connection.
     */
    username?: string | null;
    websocketConfiguration?: WebsocketConfiguration;
};
// ...
export type Adapter = {
    config?: JsonNode;
    /**
     * The adapter id, must be unique and only contain alpha numeric characters with spaces and hyphens.
     */
    id: string;
    status?: Status;
    /**
     * The adapter type associated with this instance
     */
    type?: string;
};

/**
 * The maximum MQTT-QoS for the outgoing messages.
 */
export type QoS = 'AT_MOST_ONCE' | 'AT_LEAST_ONCE' | 'EXACTLY_ONCE';

/**
 * User properties to be added to each outgoing mqtt message.
 */
export type MqttUserProperty = {
    name: string;
    value: string;
};

/**
 * List of result items that are returned by this endpoint
 */
export type NorthboundMapping = {
    /**
     * Should tag names be included when sent out.
     */
    includeTagNames?: boolean;
    /**
     * Should the timestamp be included when sent out.
     */
    includeTimestamp?: boolean;
    maxQoS?: QoS;
    /**
     * The message expiry interval.
     */
    messageExpiryInterval?: number;
    /**
     * The tag for which values hould be collected and sent out.
     */
    tagName: string;
    /**
     * The target mqtt topic where received tags should be sent to.
     */
    topic: string;
    /**
     * User properties to be added to each outgoing mqtt message.
     */
    userProperties?: Array<MqttUserProperty>;
};
// ...
export type FieldMapping = {
    /**
     * List of instructions to be applied to incoming data
     */
    instructions: Array<Instruction>;
};

/**
 * List of result items that are returned by this endpoint
 */
export type SouthboundMapping = {
    fieldMapping?: FieldMapping;
    /**
     * The tag for which values hould be collected and sent out.
     */
    tagName: string;
    /**
     * The filter defining what topics we will receive messages from.
     */
    topicFilter: string;
};

/**
 * List of result items that are returned by this endpoint
 */
export type DomainTag = {
    definition: JsonNode;
    /**
     * A user created description for this tag.
     */
    description?: string;
    /**
     * The name of the tag that identifies it within this edge instance.
     */
    name: string;
};
// ...
export type TopicFilter = {
    /**
     * The name for this topic filter.
     */
    description?: string;
    /**
     * The optional json schema for this topic filter in the data uri format.
     */
    schema?: string;
    /**
     * The topic filter according to the MQTT specification.
     */
    topicFilter: string;
};
// ...
export type EntityReferenceList = {
    items: Array<EntityReference>;
};
// ...
export type DataCombiningList = {
    items: Array<DataCombining>;
};

/**
 * A data combiner, bringing tags (adapters) and topic filters (bridges) together for further northbound data mapping
 */
export type Combiner = {
    /**
     * The unique id of the data combiner
     */
    id: string;
    /**
     * The user-facing name of the combiner
     */
    name: string;
    /**
     * The user-facing description of the combiner
     */
    description?: string;
    sources: EntityReferenceList;
    mappings: DataCombiningList;
};